Moving things around in workbox core and routing. Need to fix public and private exports
