+++ packages/workbox-core/_private/cacheWrapper.mjs
