+++ packages/workbox-core/_private/fetchWrapper.mjs
@@ -14,7 +14,7 @@
   limitations under the License.
 */
 
-import WorkboxError from '../models/WorkboxError.mjs';
+import WorkboxError from './WorkboxError.mjs';
 import '../_version.mjs';
 
 /**
