+++ packages/workbox-core/_private/indexedDBHelper.mjs
