+++ packages/workbox-core/_private/logger.mjs
