+++ packages/workbox-routing/utils/normalizeHandler.mjs
