+++ test/workbox-routing/static/sw.js
@@ -1,7 +1,9 @@
 importScripts('../../../../packages/workbox-core/build/browser/workbox-core.dev.js');
 importScripts('../../../../packages/workbox-routing/build/browser/workbox-routing.dev.js');
 
-const routing = self.workbox.routing.default;
+console.log(self.workbox.core);
+
+const routing = self.workbox.routing;
 const Route = self.workbox.routing.Route;
 
 const specialImgUrl = new URL('/test/workbox-routing/static/demo-img.png', location).toString();
